Refactory updates
