+++ src/components/characters/character-detail.tsx
@@ -1,0 +1,3 @@
+export const CharactersDetail = () => {
+    return <>teste</>;
+};
