+++ src/pages/character-detail/index.ts
