+++ src/pages/index.ts
@@ -1,4 +1,4 @@
 export * from "./home";
 export * from "./episodes";
 export * from "./location";
-export * from "./character";
+export * from "./character-detail";
